(coming soon hopefully) Items designed specifically for use with WorldEdit tools
(note: the logo looks like those logos from jellysquid and caffeinemc but this is NOT their mod i just liked the logo style from their mods and put it here)

WEwands is a mod designed for WorldEdit, but it doesn't depend on it at all. You can install it without WorldEdit without problems. This mod adds several items that are (seemingly, wait for it) useless, aren't obtainable on survival and... That's it! No big deal.

But why did I make this mod?

Well. Allow me to explain.

WorldEdit is a World Editor for Minecraft Worlds. You can select and area, replace certain blocks with certain blocks, paint with blocks, and many things useful for decoration, fast-building and terrain design. BUT! These utilities (mostly) are more comfortable to use with objects (blocks, items, tools, weapons, etc.). And the assigned objects for each thing WorldEdit does are obtainable on survival, and at the time of using the tools for what they were originally supposed to do in vanilla, it will do what it was assigned to do in WorldEdit instead. This is annoying, as eventually, you will find these objects, and you'll have to assign it to another thing and find the object again eventually. It's a loop.

With this mod, that is OVER. Well yes, this mod is not dependent on WorldEdit, but the issue written above i got inspired. These useless items that the mod adds actually have a purpose; to not have an use in a no-cheat survival world, therefore you can assign the WorldEdit utilities to these items and just not care about the assigned objects at all if you somehow manage to find them, which if you do, just throw 'em into lava or something.
This is not only for WorldEdit, though. You can use this with any mod (or plugin if you use a plugin loader on your server (which if that's the case you should want to install the mod somehow too)) with object assignments for utilities.

The idea is simple, the concept is stupid, but it somehow works. Noooo need to worry. No problem.